scrapper_cj_ecuador/requirements.txt
requests==2.32.3
pandas==2.2.2

aio-pika==9.5.6
dependency-injector==4.45.0
pydantic-settings==2.6.1
oracledb==2.5.1
boto3==1.35.36
botocore==1.35.36